Render test dashboard component (#145)
* Move renderDashboard to component file

* update lock file
