+++ packages/dashboard/package.json
@@ -42,25 +42,28 @@
     "test:cypress-ct-dev": "cypress open-ct --env disableSnapshotTests=true"
   },
   "devDependencies": {
+    "@4tw/cypress-drag-drop": "^2.2.1",
     "@cypress/vue": "^3.1.0",
     "@cypress/webpack-dev-server": "^1.8.0",
-    "@4tw/cypress-drag-drop": "^2.2.1",
     "@stencil/postcss": "^2.1.0",
     "@types/box-intersect": "^1.0.0",
     "@types/cypress-image-snapshot": "3.1.6",
+    "@vue/cli-plugin-typescript": "^5.0.7",
+    "@vue/cli-service": "^4.5.15",
+    "@vue/compiler-dom": "^3.2.37",
     "cypress": "^7.7.0",
     "cypress-image-snapshot": "4.0.1",
-    "@vue/cli-service": "^4.5.15",
     "cypress-wait-until": "^1.7.2",
     "postcss-url": "^10.1.3",
-    "rollup-plugin-node-polyfills": "^0.2.1"
+    "rollup-plugin-node-polyfills": "^0.2.1",
+    "vue": "^3.2.37"
   },
   "dependencies": {
     "@iot-app-kit/core": "^1.3.0",
     "@stencil/core": "^2.15.1",
     "@synchro-charts/core": "^4.0.0",
-    "uuid": "^8.3.2",
     "box-intersect": "^1.0.2",
-    "resize-observer-polyfill": "^1.5.1"
+    "resize-observer-polyfill": "^1.5.1",
+    "uuid": "^8.3.2"
   }
 }
